Moved lint_rules testing boilerplate to conftest.py
