+++ lint_rules/tests/test_ifs_coding_standards_2011.py
@@ -9,12 +9,13 @@ import importlib
 from pathlib import Path
 import pytest
 
-from loki import FP, HAVE_FP, Sourcefile
-from loki.lint import Reporter, Linter, DefaultHandler
+from conftest import run_linter, available_frontends
+from loki import Sourcefile
+from loki.lint import DefaultHandler
 
 
-pytestmark = pytest.mark.skipif(not HAVE_FP,
-                                reason='Fparser frontend not available')
+pytestmark = pytest.mark.skipif(not available_frontends(),
+                                reason='Suitable frontend not available')
 
 
 @pytest.fixture(scope='module', name='rules')
@@ -23,26 +24,11 @@ def fixture_rules():
     return rules
 
 
-@pytest.fixture(scope='module', name='frontend')
-def fixture_frontend():
-    """Choose frontend to use (Linter currently relies exclusively on Fparser)"""
-    return FP
-
-
-def run_linter(sourcefile, rule_list, config=None, handlers=None):
-    """
-    Run the linter for the given source file with the specified list of rules.
-    """
-    reporter = Reporter(handlers)
-    linter = Linter(reporter, rules=rule_list, config=config)
-    linter.check(sourcefile)
-    return linter
-
-
-@pytest.mark.parametrize('frontend, nesting_depth, lines', [
-    (FP, 3, []),
-    (FP, 2, [6, 12, 16, 22, 28, 35]),
-    (FP, 1, [5, 6, 10, 12, 16, 22, 27, 28, 34, 35])])
+@pytest.mark.parametrize('frontend', available_frontends())
+@pytest.mark.parametrize('nesting_depth, lines', [
+    (3, []),
+    (2, [6, 12, 16, 22, 28, 35]),
+    (1, [5, 6, 10, 12, 16, 22, 27, 28, 34, 35])])
 def test_code_body_messages(rules, frontend, nesting_depth, lines):
     '''
     Test the number and content of messages generated by CodeBodyRule
@@ -105,6 +91,7 @@ end subroutine routine_nesting
         assert f'l. {ref_line}' in msg
 
 
+@pytest.mark.parametrize('frontend', available_frontends())
 def test_module_naming(rules, frontend):
     '''Test file and modules for checking that naming is correct and matches each other.'''
     fcode = """
@@ -149,6 +136,7 @@ end module module_naming
     assert all(keyword in messages[2] for keyword in ('module_naming_mod.f90', 'filename'))
 
 
+@pytest.mark.parametrize('frontend', available_frontends())
 def test_dr_hook_okay(rules, frontend):
     fcode = """
 subroutine routine_okay
@@ -194,6 +182,7 @@ end subroutine routine_okay
     assert len(messages) == 0
 
 
+@pytest.mark.parametrize('frontend', available_frontends())
 def test_dr_hook_routine(rules, frontend):
     fcode = """
 subroutine routine_not_okay_a
@@ -315,6 +304,7 @@ end subroutine routine_not_okay_e
                for letter, i in (('a', 0), ('c', 4), ('c', 5), ('d', 6), ('e', 7)))
 
 
+@pytest.mark.parametrize('frontend', available_frontends())
 def test_dr_hook_module(rules, frontend):
     fcode = """
 module some_mod
@@ -385,10 +375,11 @@ end module some_mod
     assert '(l. 45)' in messages[3]
 
 
-@pytest.mark.parametrize('frontend, max_num_statements, passes', [
-    (FP, 10, True),
-    (FP, 4, True),
-    (FP, 3, False)])
+@pytest.mark.parametrize('frontend', available_frontends())
+@pytest.mark.parametrize('max_num_statements, passes', [
+    (10, True),
+    (4, True),
+    (3, False)])
 def test_limit_subroutine_stmts(rules, frontend, max_num_statements, passes):
     '''Test for different maximum allowed number of executable statements and
     content of messages generated by LimitSubroutineStatementsRule.'''
@@ -421,11 +412,12 @@ end subroutine routine_limit_statements
     assert all(all(keyword in msg for keyword in keywords) for msg in messages)
 
 
-@pytest.mark.parametrize('frontend, max_num_arguments, passes', [
-    (FP, 10, True),
-    (FP, 8, True),
-    (FP, 7, False),
-    (FP, 1, False)])
+@pytest.mark.parametrize('frontend', available_frontends())
+@pytest.mark.parametrize('max_num_arguments, passes', [
+    (10, True),
+    (8, True),
+    (7, False),
+    (1, False)])
 def test_max_dummy_args(rules, frontend, max_num_arguments, passes):
     '''Test for different maximum allowed number of dummy arguments and
     content of messages generated by MaxDummyArgsRule.'''
@@ -447,6 +439,7 @@ end subroutine routine_max_dummy_args
     assert all(all(keyword in msg for keyword in keywords) for msg in messages)
 
 
+@pytest.mark.parametrize('frontend', available_frontends())
 def test_mpl_cdstring(rules, frontend):
     fcode = """
 subroutine routine_okay
@@ -482,6 +475,7 @@ end subroutine routine_also_not_okay
     assert sum('(l. 18)' in msg for msg in messages) == 1
 
 
+@pytest.mark.parametrize('frontend', available_frontends())
 def test_implicit_none(rules, frontend):
     fcode = """
 subroutine routine_okay
@@ -567,6 +561,7 @@ end module mod_also_not_okay
     assert sum('"contained_contained_routine_not_okay"' in msg for msg in messages) == 1
 
 
+@pytest.mark.parametrize('frontend', available_frontends())
 def test_explicit_kind(rules, frontend):
     fcode = """
 subroutine routine_okay
@@ -624,6 +619,7 @@ end subroutine routine_not_okay
         assert all(kw in msg for kw in keys if kw is not None)
 
 
+@pytest.mark.parametrize('frontend', available_frontends())
 def test_banned_statements_default(rules, frontend):
     '''Test for banned statements with default.'''
     fcode = """
@@ -649,11 +645,12 @@ end subroutine banned_statements
     assert all(any(keyword in msg for keyword in banned_statements) for msg in messages)
 
 
-@pytest.mark.parametrize('frontend, banned_statements, passes', [
-    (FP, [], True),
-    (FP, ['GO TO'], False),
-    (FP, ['GO TO', 'RETURN'], False),
-    (FP, ['RETURN'], True)])
+@pytest.mark.parametrize('frontend', available_frontends())
+@pytest.mark.parametrize('banned_statements, passes', [
+    ([], True),
+    (['GO TO'], False),
+    (['GO TO', 'RETURN'], False),
+    (['RETURN'], True)])
 def test_banned_statements_config(rules, frontend, banned_statements, passes):
     '''Test for banned statements with custom config.'''
     fcode = """
@@ -678,6 +675,7 @@ end subroutine banned_statements
     assert all(all(keyword in msg for keyword in keywords) for msg in messages)
 
 
+@pytest.mark.parametrize('frontend', available_frontends())
 def test_fortran_90_operators(rules, frontend):
     '''Test for existence of non Fortran 90 comparison operators.'''
     fcode = """
